+++ pkg/scheduler/actions/reclaim/reclaim.go
@@ -10,7 +10,6 @@ import (
 	"github.com/NVIDIA/KAI-scheduler/pkg/scheduler/actions/common/solvers"
 	"github.com/NVIDIA/KAI-scheduler/pkg/scheduler/actions/utils"
 	"github.com/NVIDIA/KAI-scheduler/pkg/scheduler/api/common_info"
-	"github.com/NVIDIA/KAI-scheduler/pkg/scheduler/api/pod_info"
 	"github.com/NVIDIA/KAI-scheduler/pkg/scheduler/api/podgroup_info"
 	"github.com/NVIDIA/KAI-scheduler/pkg/scheduler/framework"
 	"github.com/NVIDIA/KAI-scheduler/pkg/scheduler/log"
@@ -47,7 +46,6 @@ func (ra *reclaimAction) Execute(ssn *framework.Session) {
 
 	for !jobsOrderByQueues.IsEmpty() {
 		job := jobsOrderByQueues.PopNextJob()
-
 		if !ssn.CanReclaimResources(job) {
 			continue
 		}
@@ -98,20 +96,10 @@ func (ra *reclaimAction) attemptToReclaimForSpecificJob(
 	feasibleNodes := common.FeasibleNodesForJob(maps.Values(ssn.Nodes), reclaimer)
 	solver := solvers.NewJobsSolver(
 		feasibleNodes,
-		reclaimableScenarioCheck(ssn, reclaimer),
+		ssn.ReclaimScenarioValidatorFn,
 		getOrderedVictimsQueue(ssn, reclaimer),
 		framework.Reclaim)
 	return solver.Solve(ssn, reclaimer)
-}
-
-func reclaimableScenarioCheck(ssn *framework.Session,
-	reclaimer *podgroup_info.PodGroupInfo) solvers.SolutionValidator {
-	return func(
-		_ *podgroup_info.PodGroupInfo,
-		victimJobs []*podgroup_info.PodGroupInfo,
-		victimTasks []*pod_info.PodInfo) bool {
-		return ssn.ReclaimScenarioValidator(reclaimer, victimJobs, victimTasks)
-	}
 }
 
 func getOrderedVictimsQueue(ssn *framework.Session, reclaimer *podgroup_info.PodGroupInfo) solvers.GenerateVictimsQueue {