+++ pkg/scheduler/cache/status_updater/default_status_updater.go
@@ -69,6 +69,8 @@ type defaultStatusUpdater struct {
 
 	inFlightPodGroups sync.Map
 	inFlightPods      sync.Map
+
+	appliedPodGroupUpdates sync.Map
 }
 
 // +kubebuilder:rbac:groups="",resources=events,verbs=create;update;patch;delete;list;get;watch
@@ -191,16 +193,17 @@ func (su *defaultStatusUpdater) RecordJobStatusEvent(job *podgroup_info.PodGroup
 		su.recordStaleJobEvent(job)
 	}
 
-	if !job.IsReadyForScheduling() {
-		su.recordJobNotReadyEvent(job)
-		return nil
+	updatePodgroupStatus := false
+	if job.GetNumPendingTasks() > 0 || job.GetNumGatedTasks() > 0 {
+		if !job.IsReadyForScheduling() {
+			su.recordJobNotReadyEvent(job)
+			return nil
+		}
+		if err := su.recordUnschedulablePodsEvents(job); err != nil {
+			return err
+		}
+		updatePodgroupStatus = su.recordUnschedulablePodGroup(job)
 	}
-
-	if err := su.recordUnschedulablePodsEvents(job); err != nil {
-		return err
-	}
-
-	updatePodgroupStatus := su.recordUnschedulablePodGroup(job)
 
 	if len(patchData) > 0 || updatePodgroupStatus {
 		su.pushToUpdateQueue(