+++ CHANGELOG.md
@@ -6,8 +6,23 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+## [v0.6.0] - 2025-06-16
+
+### Changed
+- Changed `runai-reservation` namespace to `kai-resource-reservation`. For migration guide, refer to this [doc](docs/migrationguides/README.md)
+- Changed `runai/queue` label key to `kai.scheduler/queue`. For migration guide, refer to [doc](docs/migrationguides/README.md)
+
 ### Fixes
 - Fixed pod status scheduled race condition between the scheduler and the pod binding
+- Removed redundant `replicas` key for binder from `values.yaml` as it is not used and not supported
+
+### Removed
+- Removed `runai-job-id` and `runai/job-id` annotations from pods and podgroups
+
+### Added
+- Added [minruntime](docs/plugins/minruntime.md) plugin, allowing PodGroups to run for a configurable amount of time without being reclaimed/preempted.
+- PodGroup Controller that will update podgroups statuses with allocation data.
+- Queue Controller that will update queues statuses with allocation data.
 
 
 ## [v0.5.1] - 2025-05-20